requirement.txt
mpmath==1.3.0
numpy==2.1.3
pillow==11.0.0
sympy==1.13.3
